output_projects/New-Project_20260329_140917/backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
python-multipart==0.0.9
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
azure-identity==1.16.0
azure-storage-blob==12.20.0
azure-keyvault-secrets==4.8.0
openai==1.30.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
redis==5.0.4
celery==5.4.0
httpx==0.27.0
pandas==2.2.2
openpyxl==3.1.2
pypdf2==3.0.1
markdownify==0.12.1
loguru==0.7.2
pydantic-settings==2.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
respx==0.21.1